fix: userChange limited only for technician and client. fix: soft delete users now can be activated and deactivated. chore: unified route for token requests

+++ backend/src/main/java/com/trackmyfix/trackmyfix/services/Impl/ClientService.java
@@ -47,9 +47,15 @@ public class ClientService implements IUserService<UserResponseDTO> {
 
     @Override
     public Map<String,String> delete(Long id) {
-        this.findById(id);
-        clientRepository.deleteById(id);
-        return Map.of("message","User id: "+id+" marked as inactive success");
+        Client client = clientRepository.findById(id).orElseThrow(()->new UserNotFoundException("User not found"));
+        if (client.getActive()){
+            clientRepository.deleteById(id);
+            return Map.of("message","User id: "+id+" marked as INACTIVE success");
+        } else {
+            client.setActive(true);
+            clientRepository.save(client);
+            return Map.of("message","User id: "+id+" marked as ACTIVE success");
+        }
     }
 
     private UserResponseDTO mapToDTO(Client user) {
